Move auto_session_state into its own module
