+++ lib/streamlit/elements/json.py
@@ -17,7 +17,7 @@ from typing import cast
 
 import streamlit
 from streamlit.proto.Json_pb2 import Json as JsonProto
-from streamlit.state.session_state import AutoSessionState
+from streamlit.state import AutoSessionState
 
 
 class JsonMixin:
