+++ lib/streamlit/state/session_state.py
@@ -15,7 +15,6 @@
 from copy import deepcopy
 import json
 from streamlit.stats import CacheStat, CacheStatsProvider
-from streamlit.type_util import Key
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -275,13 +274,6 @@ class WStates(MutableMapping[str, Any]):
 def _missing_key_error_message(key: str) -> str:
     return (
         f'st.session_state has no key "{key}". Did you forget to initialize it? '
-        f"More info: https://docs.streamlit.io/library/advanced-features/session-state#initialization"
-    )
-
-
-def _missing_attr_error_message(attr_name: str) -> str:
-    return (
-        f'st.session_state has no attribute "{attr_name}". Did you forget to initialize it? '
         f"More info: https://docs.streamlit.io/library/advanced-features/session-state#initialization"
     )
 
@@ -676,105 +668,12 @@ def _is_internal_key(key: str) -> bool:
     return key.startswith(STREAMLIT_INTERNAL_KEY_PREFIX)
 
 
-_state_use_warning_already_displayed = False
-
-
-def get_session_state() -> SessionState:
-    """Get the SessionState object for the current session.
-
-    Note that in streamlit scripts, this function should not be called
-    directly. Instead, SessionState objects should be accessed via
-    st.session_state.
-    """
-    global _state_use_warning_already_displayed
-    from streamlit.script_run_context import get_script_run_ctx
-
-    ctx = get_script_run_ctx()
-
-    # If there is no script run context because the script is run bare, have
-    # session state act as an always empty dictionary, and print a warning.
-    if ctx is None:
-        if not _state_use_warning_already_displayed:
-            _state_use_warning_already_displayed = True
-            if not st._is_running_with_streamlit:
-                logger.warning(
-                    "Session state does not function when running a script without `streamlit run`"
-                )
-        return SessionState()
-    return ctx.session_state
-
-
-class AutoSessionState(MutableMapping[str, Any]):
-    """A SessionState interface that acts as a wrapper around the
-    current script thread's SessionState instance.
-
-    When a user script uses `st.session_state`, it's interacting with
-    the singleton AutoSessionState instance, which delegates to the
-    SessionState for the active AppSession.
-
-    (This will only be used within an app script, when an AppSession is
-    guaranteed to exist.)
-    """
-
-    @staticmethod
-    def _validate_key(key: str) -> None:
-        """Raise an Exception if the given value key is invalid."""
-        if _is_widget_id(key):
-            raise StreamlitAPIException(
-                f"Keys beginning with {GENERATED_WIDGET_KEY_PREFIX} are reserved."
-            )
-
-    def __iter__(self) -> Iterator[Any]:
-        state = get_session_state()
-        return iter(state.filtered_state)
-
-    def __len__(self) -> int:
-        state = get_session_state()
-        return len(state.filtered_state)
-
-    def __str__(self) -> str:
-        state = get_session_state()
-        return str(state.filtered_state)
-
-    def __getitem__(self, key: Key) -> Any:
-        key = str(key)
-        self._validate_key(key)
-        state = get_session_state()
-        return state[key]
-
-    def __setitem__(self, key: Key, value: Any) -> None:
-        key = str(key)
-        self._validate_key(key)
-        state = get_session_state()
-        state[key] = value
-
-    def __delitem__(self, key: Key) -> None:
-        key = str(key)
-        self._validate_key(key)
-        state = get_session_state()
-        del state[key]
-
-    def __getattr__(self, key: str) -> Any:
-        self._validate_key(key)
-        try:
-            return self[key]
-        except KeyError:
-            raise AttributeError(_missing_attr_error_message(key))
-
-    def __setattr__(self, key: str, value: Any) -> None:
-        self._validate_key(key)
-        self[key] = value
-
-    def __delattr__(self, key: str) -> None:
-        self._validate_key(key)
-        try:
-            del self[key]
-        except KeyError:
-            raise AttributeError(_missing_attr_error_message(key))
-
-    def to_dict(self) -> Dict[str, Any]:
-        state = get_session_state()
-        return state.filtered_state
+def validate_key(key: str) -> None:
+    """Raise an Exception if the given value key is invalid."""
+    if _is_widget_id(key):
+        raise StreamlitAPIException(
+            f"Keys beginning with {GENERATED_WIDGET_KEY_PREFIX} are reserved."
+        )
 
 
 @attr.s(auto_attribs=True, slots=True)
